+++ StellAirium/src/StellAiriumDialog.hpp
@@ -1,12 +1,13 @@
 #pragma once
 
 #include "StelDialog.hpp"
+#include "StellAirium.hpp"
 
 class QDoubleSpinBox;
 class QSpinBox;
 class QLabel;
 class QCheckBox;
-class StellAirium;
+class QComboBox;
 
 class StellAiriumDialog : public StelDialog
 {
@@ -25,12 +26,16 @@ private slots:
     void onRadiusChanged(double km);
     void onRefreshChanged(int secs);
     void onShowGroundChanged(int state);
+    void onSourceChanged(int index);
+    void onActiveSourceChanged(StellAirium::DataSource s);
     void updateStatus();
 
 private:
     QDoubleSpinBox* radiusSpin_  {nullptr};
     QSpinBox*       refreshSpin_ {nullptr};
     QCheckBox*      groundCheck_ {nullptr};
+    QComboBox*      sourceCombo_ {nullptr};
+    QLabel*         sourceStatus_{nullptr};
     QLabel*         statusLabel_ {nullptr};
     StellAirium*    plugin_      {nullptr};
 };